Allow passing an options map to validator
I figured out that inline dereferencing is not the right default after
all. I had misunderstood it and in most cases the canonical mode does
the right thing whereas the inline mode also allows weird behavior.

Therefore, let's not make it default, but let's allow enabling it anyway
for experimenting.